Added CRUD functional on backend

+++ src/main/java/com/example/name/mapper/SubscriberMapper.java
@@ -7,13 +7,13 @@ import com.example.name.model.Subscriber;
 public interface SubscriberMapper {
     public void insert(Subscriber subscriber);
 
-    public Subscriber getById(Integer subscriberId);
+    public Subscriber getById(Long subscriberId);
 
     public List<Subscriber> getAll();
     
-    public List<Subscriber> getByFavoriteChennel(Integer ChennelId);
+    public List<Subscriber> getByFavoriteChennel(Long ChennelId);
 
     public void update(Subscriber subscriber);
 
-    public void delete(Integer subscriberId);
+    public void delete(Long subscriberId);
 }
